+++ packages/core/src/components/bearer-verifier/component.ts
@@ -5,14 +5,14 @@
 import {Binding, Component, inject, ProviderMap} from '@loopback/core';
 import {Class, Model, Repository} from '@loopback/repository';
 import {Strategies} from 'loopback4-authentication';
+import {JwtKeysRepository} from '../../repositories';
 import {ILogger, LOGGER} from '../logger-extension';
-
 import {
   BearerVerifierBindings,
   BearerVerifierConfig,
   BearerVerifierType,
 } from './keys';
-import {RevokedToken} from './models';
+import {JwtKeys, RevokedToken} from './models';
 import {FacadesBearerAsymmetricTokenVerifyProvider} from './providers/facades-bearer-asym-token-verify.provider';
 import {FacadesBearerTokenVerifyProvider} from './providers/facades-bearer-token-verify.provider';
 import {ServicesBearerAsymmetricTokenVerifyProvider} from './providers/services-bearer-asym-token-verifier';
@@ -26,9 +26,9 @@ export class BearerVerifierComponent implements Component {
     @inject(LOGGER.LOGGER_INJECT) public logger: ILogger,
   ) {
     this.providers = {};
-    this.repositories = [RevokedTokenRepository];
 
-    this.models = [RevokedToken];
+    this.repositories = [RevokedTokenRepository, JwtKeysRepository];
+    this.models = [RevokedToken, JwtKeys];
 
     if (this.config && this.config.type === BearerVerifierType.service) {
       this.providers[Strategies.Passport.BEARER_TOKEN_VERIFIER.key] =