+++ services/authentication-service/src/providers/jwt-verify.provider.ts
@@ -3,22 +3,24 @@
 // This software is released under the MIT License.
 // https://opensource.org/licenses/MIT
 import {Provider, inject} from '@loopback/core';
-import * as fs from 'fs/promises';
+import {repository} from '@loopback/repository';
 import * as jwt from 'jsonwebtoken';
+import * as jose from 'node-jose';
 import {AuthServiceBindings} from '../keys';
+import {JwtKeysRepository} from '../repositories';
 import {IAuthServiceConfig} from '../types';
 import {JWTVerifierFn} from './types';
 export class JWTAsymmetricVerifierProvider<T>
   implements Provider<JWTVerifierFn<T>>
 {
   constructor(
+    @repository(JwtKeysRepository)
+    public jwtKeysRepo: JwtKeysRepository,
     @inject(AuthServiceBindings.Config, {optional: true})
     private readonly authConfig?: IAuthServiceConfig,
   ) {}
   value(): JWTVerifierFn<T> {
     return async (code: string, options: jwt.VerifyOptions) => {
-      // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      let publicKey: any; //NOSONAR
       let payload: T;
       if (this.authConfig?.useSymmetricEncryption) {
         payload = jwt.verify(code, process.env.JWT_SECRET as string, {
@@ -27,10 +29,30 @@ export class JWTAsymmetricVerifierProvider<T>
           algorithms: ['HS256'],
         }) as T;
       } else {
-        publicKey = (await fs.readFile(
-          process.env.JWT_PUBLIC_KEY ?? '',
-        )) as Buffer;
-        payload = jwt.verify(code, publicKey, {
+        // Get the key that matches the token's kid
+        const decoded = jwt.decode(code.trim(), {complete: true});
+        if (!decoded) {
+          throw new Error('Code is not valid');
+        }
+        const kid = decoded?.header.kid;
+
+        // Load the JWKS
+        const key = await this.jwtKeysRepo.findOne({
+          where: {
+            keyId: kid,
+          },
+        });
+
+        if (!key) {
+          throw new Error('Key not found for verification');
+        }
+
+        // Convert the JWK to PEM format for verification
+        const jwkKey = await jose.JWK.asKey(key.publicKey, 'pem');
+        const pem = jwkKey.toPEM(false);
+
+        // Verify the token with the retrieved PEM-formatted public key
+        payload = jwt.verify(code, pem, {
           ...options,
           issuer: process.env.JWT_ISSUER,
           algorithms: ['RS256'],