+++ services/reporting-service/migration.js
@@ -2,14 +2,17 @@ const dotenvExt = require('dotenv-extended');
 const fs = require('fs');
 const path = require('path');
 let isLocal = false;
+
 dotenvExt.load({
   path: path.join(process.env.INIT_CWD ?? '.', '.env'),
   defaults: path.join(process.env.INIT_CWD ?? '.', '.env.defaults'),
   errorOnMissing: false,
   includeProcessEnv: true,
 });
+
 const type = 'REPORTING_SERVICE';
 
+// Check if running in a local environment
 try {
   if (fs.existsSync('.infolder')) {
     isLocal = true;
@@ -17,6 +20,7 @@ try {
 } catch (err) {
   console.info('\n');
 }
+
 if (isLocal) {
   console.info(`Skipping migrations`);
 } else if (
@@ -26,22 +30,40 @@ if (isLocal) {
     `${type}_MIGRATION or SOURCELOOP_MIGRATION variables not found in the environment, skipping automigration.`,
   );
 } else {
+  // Determine the migration type based on MYSQL_MIGRATION flag
+  const isMySQLMigration = process.env.MYSQL_MIGRATION === 'true';
+  const migrationType = isMySQLMigration ? 'mysql' : 'pg'; // Default to 'pg'
+
   const DBMigrate = require('db-migrate');
+
+  // Load the appropriate .env schema for migration type
   dotenvExt.load({
     schema: path.join('.', 'migrations', '.env.schema'),
     path: path.join(process.env.INIT_CWD, '.env'),
     errorOnMissing: true,
     includeProcessEnv: true,
   });
-  const dbmigrate = DBMigrate.getInstance(true);
+
+  // Initialize db-migrate with the correct migration directory and database config path
+  const dbmigrate = DBMigrate.getInstance(true, {
+    config: path.join(`migrations/${migrationType}`, 'database.json'),
+    cwd: path.resolve(process.cwd(), `migrations/${migrationType}`),
+  });
   dbmigrate.up();
 }
 
+// Check if migration copy is needed and copy migration files accordingly
 if (
   process.env.SOURCELOOP_MIGRATION_COPY ||
   process.env[`${type}_MIGRATION_COPY`]
 ) {
-  copyFolderRecursiveSync(path.join('.', 'migrations'), process.env.INIT_CWD);
+  const isMySQLMigration = process.env.MYSQL_MIGRATION === 'true';
+  const migrationType = isMySQLMigration ? 'mysql' : 'pg'; // Default to 'pg'
+
+  copyFolderRecursiveSync(
+    path.join('.', 'migrations', migrationType, 'migrations'),
+    process.env.INIT_CWD,
+  );
 }
 
 function copyFileSync(source, target) {