+++ services/task-service/src/__tests__/fixtures/test-helper.ts
@@ -8,11 +8,9 @@ import {
 } from '@loopback/testlab';
 import {AuthCacheSourceName, ILogger, LOGGER} from '@sourceloop/core';
 import {Task} from 'camunda-external-task-client-js';
-import * as fs from 'fs';
-import {sign} from 'jsonwebtoken';
 import {AuthenticationBindings} from 'loopback4-authentication';
-import path from 'path';
 import {TaskDbSourceName, WorkflowServiceSourceName} from '../../types';
+import {AuthenticationDbDataSource} from './datasources';
 import {TestTaskServiceApplication} from './test.application';
 
 export async function setupApplication(loggerStub?: ILogger) {
@@ -33,6 +31,12 @@ export async function setupApplication(loggerStub?: ILogger) {
       connector: 'memory',
     }),
   );
+
+  app.bind('datasources.AuthDB').to({
+    name: 'AuthDB',
+    connector: 'memory',
+  });
+  app.dataSource(AuthenticationDbDataSource);
 
   app.bind(`datasources.config.${AuthCacheSourceName}`).to({
     name: 'redis',
@@ -79,26 +83,6 @@ export async function clearRepo<T extends {deleteAllHard: () => Promise<void>}>(
   await repo.deleteAllHard();
 }
 
-export function getToken(permissions: string[] = []) {
-  process.env.JWT_PRIVATE_KEY = path.resolve(
-    __dirname,
-    '../../../src/__tests__/unit/utils/privateKey.txt',
-  );
-  const privateKey = fs.readFileSync(process.env.JWT_PRIVATE_KEY ?? '');
-  return `Bearer ${sign(
-    {
-      id: 'test',
-      userTenantId: 'test',
-      iss: process.env.JWT_ISSUER,
-      permissions,
-    },
-    privateKey,
-    {
-      algorithm: 'RS256',
-    },
-  )}`;
-}
-
 export function getMockCamundaParameters(data: AnyObject) {
   const stubTaskService = {
     complete: sinon.stub(),
@@ -119,12 +103,4 @@ export function getMockCamundaParameters(data: AnyObject) {
 
 function setupEnv() {
   process.env.JWT_ISSUER = 'test';
-  process.env.JWT_PRIVATE_KEY = path.resolve(
-    __dirname,
-    '../../../src/__tests__/unit/utils/privateKey.txt',
-  );
-  process.env.JWT_PUBLIC_KEY = path.resolve(
-    __dirname,
-    '../../../src/__tests__/unit/utils/publicKey.txt',
-  );
 }