+++ sandbox/notification-socket-example/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## <small>6.0.24 (2024-12-05)</small>
+
+**Note:** Version bump only for package @sourceloop/notification-socket-example
+
+
+
+
+
+## <small>6.0.23 (2024-11-22)</small>
+
+**Note:** Version bump only for package @sourceloop/notification-socket-example
+
+
+
+
+
 ## <small>6.0.22 (2024-10-22)</small>
 
 **Note:** Version bump only for package @sourceloop/notification-socket-example