+++ sandbox/oauth-example/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## <small>0.5.11 (2024-12-05)</small>
+
+**Note:** Version bump only for package @sourceloop/oauth-example-api
+
+
+
+
+
+## <small>0.5.10 (2024-11-22)</small>
+
+**Note:** Version bump only for package @sourceloop/oauth-example-api
+
+
+
+
+
 ## <small>0.5.9 (2024-10-22)</small>
 
 **Note:** Version bump only for package @sourceloop/oauth-example-api