+++ sandbox/search-ms-example/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## <small>0.9.5 (2024-12-05)</small>
+
+**Note:** Version bump only for package @sourceloop/search-ms-example
+
+
+
+
+
+## <small>0.9.4 (2024-11-22)</small>
+
+**Note:** Version bump only for package @sourceloop/search-ms-example
+
+
+
+
+
 ## <small>0.9.3 (2024-10-22)</small>
 
 **Note:** Version bump only for package @sourceloop/search-ms-example